added Case Report #TASK-782
